+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add `init` alias for `config:init` command [TPW-56]
+- Add interactive shell auto-completion setup prompt at the end of `config:init` command [TPW-56]
+- Automatically detect user's shell (bash or zsh) and provide installation instructions for shell completion [TPW-56]
+
 ## [2.1.5] - 2025-11-13
 
 ## [2.1.4] - 2025-11-13